fix ben name in incentive

+++ src/main/java/com/iemr/flw/repo/identity/BeneficiaryRepo.java
@@ -23,7 +23,7 @@ public interface BeneficiaryRepo extends JpaRepository<RMNCHBeneficiaryDetailsRm
     Optional<RMNCHBeneficiaryDetailsRmnch> findById(Long benID);
 
     @Query(value = "SELECT BeneficiaryRegID FROM db_identity.i_beneficiarydetails_rmnch WHERE BeneficiaryId = :benId", nativeQuery = true)
-    BigInteger getBenRegIdFromBenId(@Param("benId") Long benId);
+    Long getBenRegIdFromBenId(@Param("benId") Long benId);
 
     @Query(nativeQuery = true, value = " SELECT userid FROM db_iemr.m_user WHERE UserName = :userName ")
     Integer getUserIDByUserName(@Param("userName") String userName);
